Resolve "Revise the domain model according to the design review"

Closes #53

See merge request usi-si-teaching/bachelor-inf/2026/software-atelier-4/team-4-coffeebeans/backend!42

+++ src/main/java/ch/usi/inf/bsc/sa4/lab02spring/controller/dto/UserProfileDTO.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public record UserProfileDTO(
         String name,
-        int playedLevelsCount,
-        int completedLevelsCount,
+        long playedLevelsCount,
+        long completedLevelsCount,
         List<LevelDTO> createdLevels // List is better than Set because it preserves order for the frontend
 ) {
     /// Constructs a UserProfileDTO from the given User entity and statistics.
@@ -15,7 +15,7 @@ public record UserProfileDTO(
     /// @param playedLevelsCount the number of distinct levels the user has played
     /// @param completedLevelsCount the number of distinct levels the user has completed
     /// @param createdLevels the list of levels created by the user as LevelDTOs
-    public UserProfileDTO(User user, int playedLevelsCount, int completedLevelsCount, List<LevelDTO> createdLevels) {
+    public UserProfileDTO(User user, long playedLevelsCount, long completedLevelsCount, List<LevelDTO> createdLevels) {
         this(
                 user.getName(),
                 playedLevelsCount,
